+++ modules/Students/Letters.php
@@ -19,6 +19,8 @@ if ( $_REQUEST['modfunc'] === 'save'
 
 	$_REQUEST['mailing_labels'] = issetVal( $_REQUEST['mailing_labels'], '' );
 
+	$_REQUEST['hide_headers'] = issetVal( $_REQUEST['hide_headers'], '' );
+
 	$st_list = "'" . implode( "','", $_REQUEST['st_arr'] ) . "'";
 
 	$extra['WHERE'] = " AND s.STUDENT_ID IN (" . $st_list . ")";
@@ -115,19 +117,24 @@ if ( $_REQUEST['modfunc'] === 'save'
 	{
 		unset( $_ROSARIO['DrawHeader'] );
 
-		if ( $_REQUEST['mailing_labels'] == 'Y' )
+		if ( $_REQUEST['mailing_labels'] === 'Y' )
 		{
 			echo '<br /><br /><br />';
 		}
 
-		//DrawHeader(ParseMLField(Config('TITLE')).' Letter');
-		DrawHeader( '&nbsp;' );
-		DrawHeader( $student['FULL_NAME'], $student['STUDENT_ID'] );
-		DrawHeader( $student['GRADE_ID'], $student['SCHOOL_TITLE'] );
-		//DrawHeader('',GetMP(GetCurrentMP('QTR',DBDate(),false)));
-		DrawHeader( ProperDate( DBDate() ) );
+		if ( ! $_REQUEST['hide_headers'] )
+		{
+			DrawHeader( '&nbsp;' );
+			DrawHeader( $student['FULL_NAME'], $student['STUDENT_ID'] );
+			DrawHeader( $student['GRADE_ID'], $student['SCHOOL_TITLE'] );
+			DrawHeader( ProperDate( DBDate() ) );
+		}
+		elseif ( $_REQUEST['mailing_labels'] === 'Y' )
+		{
+			echo '<br /><br /><br /><br />';
+		}
 
-		if ( $_REQUEST['mailing_labels'] == 'Y' )
+		if ( $_REQUEST['mailing_labels'] === 'Y' )
 		{
 			echo '<br /><br /><table class="width-100p"><tr><td style="width:50px;"> &nbsp; </td><td>' . $student['MAILING_LABEL'] . '</td></tr></table><br />';
 		}
@@ -172,7 +179,7 @@ if ( ! $_REQUEST['modfunc'] )
 		$extra['search'] = '';
 
 		// FJ add TinyMCE to the textarea.
-		$extra['extra_header_left'] .= '<table class="width-100p"><tr><td>' .
+		$extra['extra_header_left'] .= '<table class="width-100p"><tr class="st"><td>' .
 		TinyMCEInput(
 			GetTemplate(),
 			'letter_text',
@@ -202,9 +209,18 @@ if ( ! $_REQUEST['modfunc'] )
 
 		$substitutions += SubstitutionsCustomFields( 'student' );
 
-		$extra['extra_header_left'] .= '<table><tr class="st"><td class="valign-top">' .
+		$extra['extra_header_left'] .= '<tr class="st"><td>' .
 			SubstitutionsInput( $substitutions ) .
-		'</td></tr></table>';
+		'</td></tr>';
+
+		// @since 11.5 Add Hide Headers option
+		$extra['extra_header_left'] .= '<tr class="st"><td>' . CheckboxInput(
+			issetVal( $_REQUEST['hide_headers'], '' ),
+			'hide_headers',
+			_( 'Hide Headers' ),
+			'',
+			true
+		) . '</td></tr></table>';
 
 		/**
 		 * Print Letters header